+++ spec/zakuro/version/all/abstract/medieval_line.rb
@@ -6,47 +6,44 @@ require_relative './medieval_month'
 module Zakuro
   # :nodoc:
   module All
-    # :nodoc:
-    module Gihou
-      # MedievalLine 行データ（中世）
-      class MedievalLine
-        # @return [Integer] 行番号
-        attr_reader :num
-        # @return [Month] 月
-        attr_reader :month
+    # MedievalLine 行データ（中世）
+    class MedievalLine
+      # @return [Integer] 行番号
+      attr_reader :num
+      # @return [Month] 月
+      attr_reader :month
 
-        #
-        # 初期化
-        #
-        # @param [Integer] num 行番号
-        # @param [Month] month 月
-        #
-        def initialize(num: -1, month: MedievalMonth.new)
-          @num = num
-          @month = month
-        end
+      #
+      # 初期化
+      #
+      # @param [Integer] num 行番号
+      # @param [Month] month 月
+      #
+      def initialize(num: -1, month: MedievalMonth.new)
+        @num = num
+        @month = month
+      end
 
-        #
-        # 不正か
-        #
-        # @return [True] 不正
-        # @return [False] 不正なし
-        #
-        def invalid?
-          @num == -1
-        end
+      #
+      # 不正か
+      #
+      # @return [True] 不正
+      # @return [False] 不正なし
+      #
+      def invalid?
+        @num == -1
+      end
 
-        #
-        # ハッシュ化する
-        #
-        # @return [Hash<String, Objcet>] ハッシュ
-        #
-        def to_h
-          {
-            num: num,
-            month: month.to_h
-          }
-        end
+      #
+      # ハッシュ化する
+      #
+      # @return [Hash<String, Objcet>] ハッシュ
+      #
+      def to_h
+        {
+          num: num,
+          month: month.to_h
+        }
       end
     end
   end
